osg_dj/oldschoolgym/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
Pillow==9.4.0
pycodestyle==2.10.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
